Fix 에러 처리 로직 수정
에러 출력 형식 통일
모든 에러는 콘솔에 출력하고, client 에는 간략한 메세지와 status 만 전달
