+++ package.json
@@ -17,19 +17,15 @@
   },
   "dependencies": {
     "aws-sdk": "^2.348.0",
-    "cookie-parser": "~1.4.4",
     "debug": "~2.6.9",
     "dotenv": "^16.0.3",
     "express": "~4.16.1",
-    "express-session": "^1.17.3",
     "firebase-admin": "^11.2.0",
-    "http-errors": "~1.6.3",
     "mongoose": "^6.6.5",
     "mongoose-autopopulate": "^0.16.1",
     "morgan": "~1.9.1",
     "multer": "^1.4.5-lts.1",
     "multer-s3-v2": "^2.2.1",
-    "pug": "2.0.0-beta11",
     "socket.io": "^4.5.3"
   },
   "devDependencies": {
