+++ routes/controllers/mindMapController.js
@@ -1,81 +1,84 @@
 const endOfMindMapListReq = async (req, res, next) => {
-  if (res.locals.mindMapsList) {
-    const responseBody = {};
+  if (!res.locals.mindMapsList) {
+    const error = new Error(
+      'No data delivered to endOfMindMapListReq in mindMapController.js',
+    );
+    error.status = 500;
 
-    responseBody.result = 'ok';
-    responseBody.mindMap = res.locals.mindMapsList;
-    responseBody.count = res.locals.mindMapsList.length;
-
-    res.status(200).json(responseBody);
-    return;
+    next(error);
   }
 
-  const error = new Error(
-    'Error : no list transferred in mindMapController.js',
-  );
-  error.status = 500;
-  next(error);
+  const responseBody = {};
+
+  responseBody.result = 'ok';
+  responseBody.mindMap = res.locals.mindMapsList;
+  responseBody.count = res.locals.mindMapsList.length;
+
+  res.status(200).json(responseBody);
 };
 
 const endOfMindMapReq = (req, res, next) => {
-  try {
-    const responseBody = {
-      result: 'ok',
-      mindMap: res.locals.mindMap,
-    };
+  if (!res.locals.mindMap) {
+    const error = new Error(
+      'No data delivered to endOfMindMapReq in mindMapController.js',
+    );
+    error.status = 500;
 
-    res.status(200).json(responseBody);
-  } catch (error) {
-    console.error(error);
     next(error);
   }
+
+  const responseBody = {
+    result: 'ok',
+    mindMap: res.locals.mindMap,
+  };
+
+  res.status(200).json(responseBody);
 };
 
 const endOfPostMindMapReq = (req, res, next) => {
-  try {
-    const responseBody = {
-      result: 'ok',
-      mindMap: res.locals.mindMap,
-      node: res.locals.childNode,
-    };
+  if (!res.locals.mindMap || !res.locals.childNode) {
+    const error = new Error(
+      'No data delivered to endOfPostMindMapReq in mindMapController.js',
+    );
+    error.status = 500;
 
-    res.status(200).json(responseBody);
-  } catch (error) {
-    console.error(error);
     next(error);
   }
+
+  const responseBody = {
+    result: 'ok',
+    mindMap: res.locals.mindMap,
+    node: res.locals.childNode,
+  };
+
+  res.status(200).json(responseBody);
 };
 
-const endOfDeleteMindMapReq = (req, res, next) => {
-  try {
-    const response = { result: 'ok' };
-    res.json(response);
-  } catch (error) {
-    console.error(error);
-    next(error);
-  }
+const endOfDeleteMindMapReq = (req, res) => {
+  const response = { result: 'ok' };
+  res.json(response);
 };
 
 const endOfMindMapAccessReq = (req, res, next) => {
-  try {
-    const response = {
-      result: 'ok',
-      access: res.locals.access,
-    };
+  if (!res.locals.access || !res.locals.mindMap) {
+    const error = new Error(
+      'No data delivered to endOfMindMapAccessReq in mindMapController.js',
+    );
+    error.status = 500;
 
-    if (res.locals.access === 'public') {
-      response.mindMap = res.locals.mindMap;
-    }
-
-    if (req.user) {
-      response.mindMap = res.locals.mindMap;
-    }
-
-    res.status(200).json(response);
-  } catch (error) {
-    console.error(error);
     next(error);
   }
+
+  const response = {
+    result: 'ok',
+    access: res.locals.access,
+  };
+
+  if (res.locals.access === 'public' || req.user) {
+    response.mindMap = res.locals.mindMap;
+  }
+
+  res.status(200).json(response);
 };
 
 module.exports = {
