+++ routes/middlewares/authMiddleware.js
@@ -2,21 +2,26 @@ const admin = require('../../configs/firebaseConfig');
 
 const auth = async (req, res, next) => {
   try {
+    if (!req.headers || typeof req.headers.authorization !== 'string') {
+      throw new Error('Invalid token delivered!!');
+    }
+
     const token = req.headers.authorization.split(' ')[1];
 
     const decodeValue = await admin.auth().verifyIdToken(token);
 
-    if (decodeValue) {
-      req.user = decodeValue;
-      return next();
+    if (!decodeValue) {
+      throw new Error('Authentication Error');
     }
 
-    throw new Error('Authentication Error');
-  } catch (error) {
-    error.status = 401;
-    error.message = 'Authentication Error';
+    req.user = decodeValue;
 
-    return next(error);
+    next();
+  } catch (error) {
+    error.message = `Error in auth in authMiddleware.js : ${error.message}`;
+    error.status = 401;
+
+    next(error);
   }
 };
