+++ routes/middlewares/authWithoutErrorMiddleware.js
@@ -2,19 +2,22 @@ const admin = require('../../configs/firebaseConfig');
 
 const authWithoutError = async (req, res, next) => {
   try {
-    const token = req.headers.authorization.split(' ')[1];
+    if (req.headers && typeof req.headers.authorization === 'string') {
+      const token = req.headers.authorization.split(' ')[1];
 
-    const decodeValue = await admin.auth().verifyIdToken(token);
+      const decodeValue = await admin.auth().verifyIdToken(token);
+      if (!decodeValue) {
+        throw new Error('Verification failed!!');
+      }
 
-    if (decodeValue) {
       req.user = decodeValue;
-      return next();
     }
 
-    throw new Error('Authentication Error');
+    next();
   } catch (error) {
-    req.user = null;
-    return next();
+    error.message = `Error in authWithoutError in authWithoutErrorMiddleware.js : ${error.message}`;
+
+    next(error);
   }
 };
