+++ routes/userRouter.js
@@ -3,7 +3,9 @@ const express = require('express');
 const router = express.Router();
 
 const auth = require('./middlewares/authMiddleware');
-const { getMyMindMapList } = require('./middlewares/dataHandlingMiddleware');
+const {
+  getMyMindMapList,
+} = require('./middlewares/mindMapDataHandlingMiddleware');
 const { endOfMindMapListReq } = require('./controllers/mindMapController');
 const { login } = require('./controllers/authController');
